backend/requirements.txt
alembic==1.12.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
asyncpg==0.29.0
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
fastapi==0.104.1
freezegun==1.2.2
greenlet==3.0.1
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
idna==3.4
iniconfig==2.0.0
Jinja2==3.1.2
lxml==4.9.3
Mako==1.3.0
MarkupSafe==2.1.3
packaging==23.2
pluggy==1.3.0
psycopg2-binary==2.9.9
pydantic==2.5.1
pydantic_core==2.14.3
pytest==7.4.3
python-dateutil==2.8.2
requests==2.31.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.23
SQLAlchemy-Utils==0.41.1
starlette==0.27.0
typing_extensions==4.8.0
urllib3==2.1.0
uvicorn==0.24.0.post1
